+++ src/main.js
@@ -129,6 +129,8 @@ ipcMain.handle('device:open-link', (_event, url) => {
 ipcMain.handle('effects:get-state', (_event, output) => effects.getState(output));
 ipcMain.handle('effects:set', (_event, id, changes, output) => effects.setEffect(id, changes || {}, output));
 ipcMain.handle('effects:set-master', (_event, enabled, output) => effects.setMaster(Boolean(enabled), output));
+ipcMain.handle('eq:get-state', (_event, output) => effects.getEqState(output));
+ipcMain.handle('eq:set', (_event, changes, output) => effects.setEq(changes || {}, output));
 
 ipcMain.handle('mic:get-state', () => capture.getState());
 ipcMain.handle('mic:set-volume', (_event, volume) => capture.setVolume(volume));